MH-10: fixed bug with skipping some
found an interesting but with the site I was working with; where some excercise types are strored in type.html and some in just type so It skipped them the 1st time; managed to scrape after that.

There was another random minor issue where what they name 'upper-back' is stored in 'middle-back' lol.

Web scrape is a success!